backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
langchain-openai
langchain
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pytz